RPC/requirements.txt
customtkinter==5.2.2
python-dotenv==1.0.1
pillow==10.4.0